Extract Ctrl+C detection to main.rs
